+++ R/RcppExports.R
@@ -5,8 +5,8 @@ force_materialization <- function(x) {
     invisible(.Call(`_vroom_force_materialization`, x))
 }
 
-vroom_materialize <- function(x) {
-    invisible(.Call(`_vroom_vroom_materialize`, x))
+vroom_materialize <- function(x, replace = FALSE) {
+    invisible(.Call(`_vroom_vroom_materialize`, x, replace))
 }
 
 vroom_str_ <- function(x) {